+++ dune/stuff/playground/grid/provider/starcd.hh
@@ -59,7 +59,7 @@ public:
   typedef GridProviderStarCD<GridType> ThisType;
 
   //! Dimension of the provided grid.
-  static const unsigned int dim = BaseType::dim;
+  static const size_t dim = BaseType::dim;
 
   //! Type of the grids coordinates.
   typedef typename BaseType::CoordinateType CoordinateType;
